fix: TestClass naming
